+++ app/src/main/java/com/ekspensify/app/data/model/responseModel/CategoryInsightsResponseModel.kt
@@ -10,5 +10,5 @@ data class CategoryInsightsResponseModel(
     var category: CategoryResponseModel? = null,
     @SerializedName("amount")
     @Expose
-    var amount: Int? = 0
+    var amount: Double? = 0.0
 ) : Serializable
